requirements.txt
Flask>=2.3.0
TensorFlow>=2.12.0
numpy>=1.23.0
Pillow>=9.0.0
opencv-python>=4.7.0
scikit-learn>=1.2.0